+++ R/mutators.R
@@ -24,6 +24,38 @@ get_order <- function(x, margin = 1) {
   ord
 }
 
+## Supplementary variables -----------------------------------------------------
+is_supplementary <- function(x, margin = 1) {
+  margin <- margin[[1L]]
+  if (margin == 1) supp <- x@rows@supplement
+  if (margin == 2) supp <- x@columns@supplement
+  supp
+}
+
+has_supplementary <- function(x, margin = 1) {
+  any(is_supplementary(x, margin = margin))
+}
+
+get_extra <- function(x) {
+  as.data.frame(x@extra)
+}
+
+has_extra <- function(x) {
+  all(lengths(x@extra) > 0)
+}
+
+`set_extra<-` <- function(x, value) {
+  ## /!\ Reorder, see build_results() /!\
+  value <- lapply(
+    X = value,
+    FUN = function(val, i) { val[i] },
+    i = get_order(x, margin = 1)
+  )
+  x@extra <- value
+  methods::validObject(x)
+  x
+}
+
 # Groups =======================================================================
 get_groups <- function(x, margin = 1) {
   margin <- margin[[1L]]